+++ bundle-polkadot-api-contract.js
@@ -39,6 +39,9 @@
     }));
   }
   function v0ToV1(registry, v0) {
+    if (!v0.metadataVersion.length) {
+      throw new Error('Invalid format for V0 (detected) contract metadata');
+    }
     return registry.createType('ContractMetadataV1', util.objectSpread({}, v0, {
       spec: util.objectSpread({}, v0.spec, {
         constructors: v0ToV1Names(v0.spec.constructors),
@@ -84,17 +87,22 @@
     }));
   }
 
-  const enumVersions = ['V3', 'V2', 'V1'];
+  function v3ToV4(registry, v3) {
+    return v3;
+  }
+
+  const enumVersions = ['V4', 'V3', 'V2', 'V1'];
   function createConverter(next, step) {
     return (registry, input) => next(registry, step(registry, input));
   }
-  function v3ToLatest(registry, v3) {
-    return v3;
+  function v4ToLatest(registry, v4) {
+    return v4;
   }
+  const v3ToLatest = createConverter(v4ToLatest, v3ToV4);
   const v2ToLatest = createConverter(v3ToLatest, v2ToV3);
   const v1ToLatest = createConverter(v2ToLatest, v1ToV2);
   const v0ToLatest = createConverter(v1ToLatest, v0ToV1);
-  const convertVersions = [['V3', v3ToLatest], ['V2', v2ToLatest], ['V1', v1ToLatest], ['V0', v0ToLatest]];
+  const convertVersions = [['V4', v4ToLatest], ['V3', v3ToLatest], ['V2', v2ToLatest], ['V1', v1ToLatest], ['V0', v0ToLatest]];
 
   const l$1 = util.logger('Abi');
   const PRIMITIVE_ALWAYS = ['AccountId', 'AccountIndex', 'Address', 'Balance'];
@@ -106,8 +114,14 @@
   }
   function getLatestMeta(registry, json) {
     const vx = enumVersions.find(v => util.isObject(json[v]));
+    const jsonVersion = json.version;
+    if (!vx && jsonVersion && !enumVersions.find(v => v === `V${jsonVersion}`)) {
+      throw new Error(`Unable to handle version ${jsonVersion}`);
+    }
     const metadata = registry.createType('ContractMetadata', vx ? {
       [vx]: json[vx]
+    } : jsonVersion ? {
+      [`V${jsonVersion}`]: json
     } : {
       V0: json
     });
@@ -279,7 +293,7 @@
     name: '@polkadot/api-contract',
     path: (({ url: (typeof document === 'undefined' && typeof location === 'undefined' ? new (require('u' + 'rl').URL)('file:' + __filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-api-contract.js', document.baseURI).href)) }) && (typeof document === 'undefined' && typeof location === 'undefined' ? new (require('u' + 'rl').URL)('file:' + __filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-api-contract.js', document.baseURI).href))) ? new URL((typeof document === 'undefined' && typeof location === 'undefined' ? new (require('u' + 'rl').URL)('file:' + __filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-api-contract.js', document.baseURI).href))).pathname.substring(0, new URL((typeof document === 'undefined' && typeof location === 'undefined' ? new (require('u' + 'rl').URL)('file:' + __filename).href : typeof document === 'undefined' ? location.href : (document.currentScript && document.currentScript.src || new URL('bundle-polkadot-api-contract.js', document.baseURI).href))).pathname.lastIndexOf('/') + 1) : 'auto',
     type: 'esm',
-    version: '9.2.4'
+    version: '9.3.1'
   };
 
   function applyOnEvent(result, types, fn) {