+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 10.1.7
 - api, api-contract, types 9.2.4
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 2.9.7
 - hw-ledger, keyring, util, util-crypto 10.1.6